Refactor insert() function and fix hash2() bug
The insert function was duplicating so many lines
of codes for each probe type when the only difference
in the code was the method to recalculate the index
to which the hash object is inserted. So the common
lines of code were moved into one while loop and the
code to recalculate index was moved to an if/else based
on the probe type (linear/double hash).

This bug messed me up for hours. The hash2 function is
the auxiliary function used in double hashing. Within it,
the modded hashCode of the key is checked if is negative
because hashCode() can possibly return a negative number.
In the case of a negative hashCode value, the tablesize is
added to the value. BUT, since the hash2 function mods the
hashCode value by tableSize - 2, the negative value should
have tableSize - 2 added to it. I was added tableSize and
that was giving me multiple problems.